Adding some hacks to handle content
The Mystery Microbes content features some content layers that are
outside of the scope of the code. So I'm having to add some messy hacks
to get the content in here. This works for now, but we'll probably look
at how to re-organize the content once we approach this fresh with the
SMM lab companions.